pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fbqc-snspe"
version = "0.1.0"
description = "Design toolkit for polarization-flattened waveguide-integrated SNSPDs"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "Federico Santona" }
]
keywords = ["SNSPD", "photonics", "quantum", "waveguide"]
license = { text = "MIT" }
dependencies = [
  "numpy>=1.24",
  "scipy>=1.10",
  "pandas>=2.0",
  "matplotlib>=3.7",
  "pydantic>=1.10",
  "pydantic<3",
  "pyyaml>=6.0",
  "jinja2>=3.1",
  "tqdm>=4.65",
  "EMpy",
  "meep>=1.25",
]

[tool.setuptools]
packages = ["fbqc_snspe"]
include-package-data = true

[tool.setuptools.package-data]
fbqc_snspe = ["../data/materials/*.csv"]
